// store/application/types/enums/getter-type.ts
/**
 * Getters supported by application store module
 */
 export enum GetterType {
    GET_IMAGE_URL_TEMPLATE = "GET_IMAGE_URL_TEMPLATE",
    GET_IMAGE_WORKAROUND_URL_TEMPLATE = "GET_IMAGE_WORKAROUND_URL_TEMPLATE",
    GET_FOCUSED_ELEMENTS = "GET_FOCUSED_ELEMENTS",
    IS_MODAL_EXIT_APP_SHOW = "IS_MODAL_EXIT_APP_SHOW",
    IS_MODAL_SIGN_OUT_SHOW = "IS_MODAL_SIGN_OUT_SHOW",
    GET_HEADER_VISIBILITY = "GET_HEADER_VISIBILITY",
    GET_SELECTED_ID_PROGRAM = "GET_SELECTED_ID_PROGRAM",
    GET_CURRENT_SELECTED_ID_PROGRAM = "GET_CURRENT_SELECTED_ID_PROGRAM",
    GET_IS_HAVE_FOCUSED_ELEMENT = "GET_IS_HAVE_FOCUSED_ELEMENT",
    IS_PREVIEW_SIGN_IN_MODAL = "IS_PREVIEW_SIGN_IN_MODAL",
    IS_SHOWING_INTERNET_MODAL = "IS_SHOWING_INTERNET_MODAL",
    GET_COUNTER_VOD = "GET_COUNTER_VOD",
    GET_COUNTER_MYTV = "GET_COUNTER_MYTV",
    GET_COUNTER_SEARCH = "GET_COUNTER_SEARCH",
    GET_COUNTER_HOME = "GET_COUNTER_HOME",
    GET_COUNTER_GENRE = "GET_COUNTER_GENRE",
    GET_COUNTER_DETAIL = "GET_COUNTER_DETAIL",
    GET_DISABLE_ROUTE_CHANGE = "GET_DISABLE_ROUTE_CHANGE",
    GET_TOOLTIPE_SHOWING = "GET_TOOLTIPE_SHOWING",
    GET_TOOLTIPE_MESSAGE = "GET_TOOLTIPE_MESSAGE",
    GET_PRIVACY_POLICY_TEMPLATE = "GET_PRIVACY_POLICY_TEMPLATE",
    GET_ABOUT_US_TEMPLATE = "GET_ABOUT_US_TEMPLATE",
    GET_VISIBLE_FORGOT = "GET_VISIBLE_FORGOT",
    GET_IS_DISABLED_SENTRY = "GET_IS_DISABLED_SENTRY",
    GET_VISIBLE_DESCRIPTION = "GET_VISIBLE_DESCRIPTION",
    GET_CURRENT_TIME = "GET_CURRENT_TIME",
    GET_FORGOT_LINK = "GET_FORGOT_LINK",
    GET_VOLUME = "GET_VOLUME",
    GET_SHOWING_TOOLTIPE_TIME = "GET_SHOWING_TOOLTIPE_TIME",
    GET_FILTER = "GET_FILTER",
    IS_MODAL_EXPIRED_CODE = "IS_MODAL_EXPIRED_CODE",
    IS_MODAL_SIGN_IN_UNAVAILABLE = "IS_MODAL_SIGN_IN_UNAVAILABLE",
    GET_FILTER_PARAMS = "GET_FILTER_PARAMS",
    IS_OKTA_TOKEN_ERROR_MODAL = "IS_OKTA_TOKEN_ERROR_MODAL",
    IS_OKTA_SERVER_ERROR_MODAL = "IS_OKTA_SERVER_ERROR_MODAL"
}